Added Generator for Crash Twinsanity and Around the Clock at Bikini Bottom
Added a variant for Crash Twinsanity that does not have hover skip-related goals. This skip is a major reset point and cards are typically reset to avoid having to do this skip.
New generator for Around the Clock at Bikini Bottom added. Some runners of the game were interested in a bingo so I added one for them.