Added Crash Twinsanity variant and new board for Around the Clock. (#178)
* Added Generator for Crash Twinsanity and Around the Clock at Bikini Bottom

Added a variant for Crash Twinsanity that does not have hover skip-related goals. This skip is a major reset point and cards are typically reset to avoid having to do this skip.
New generator for Around the Clock at Bikini Bottom added. Some runners of the game were interested in a bingo so I added one for them.

* Delete around_the_clock_at_bikini_bottom_generator.js

* Delete crash_twinsanity_hoverless_variant_generator.js

* Added hoverless variant for Crash Twinsanity

Crash Twinsanity generator without goals that require hover skip. Most bingo involves resetting until there are no cavern goals, so made a board that just doesn't have them.

* Added generator for Around the Clock at Bikini Bottom

Added generator for Around the Clock at Bikini Bottom. Members of the Discord wanted one so here it is.

* Update game_type.py

Updated to include Crash Twinsanity variant and Around the Clock.

* Update game_type.py

Co-authored-by: Kyle Buzsaki <kbuzsaki@gmail.com>