Update zelda_wind_waker_randomizer_generator.js
There was a semicolon at the end but as far as I know that shouldn't be there so I deleted it. Hopefully this works I swapped it to an SRL Format. (I have tested the bingo of course and it works for me, but I've never done a pull request before).